+++ js/jquery/components/Tabs.js
@@ -23,7 +23,7 @@
         this.nav = element.find(options.navElement);
 
         // Collection of content sections
-        this.sections = element.find(options.sectionsElement).conceal();
+        this.sections = element.find(options.sectionElement).conceal();
 
         /** Collection of tabs (anchor links) */
         this.tabs = this.nav.find('ul > li > a').each(function(index) {
@@ -51,7 +51,7 @@
         // Trigger default tab to display
         var index = options.defaultIndex;
 
-        if (options.persistState) {
+        if (options.persistState && options.cookie) {
             index = $.cookie('toolkit.tabs.' + options.cookie);
 
         } else if (options.loadFragment && location.hash) {
@@ -98,34 +98,35 @@
 
             var index = tab.data('index'),
                 section = this.sections.item(index),
-                url = this.readValue(tab, this.options.getUrl),
-                options = this.options;
+                options = this.options,
+                url = this.readValue(tab, options.getUrl),
+                isPrefix = Toolkit.options.isPrefix;
 
             // Load content with AJAX
             if (options.ajax && url && url.substr(0, 1) !== '#' && !this.cache[url]) {
                 this.requestData(url,
                     function() {
                         section.html(this._loadingTemplate())
-                            .addClass(Toolkit.options.isPrefix + 'loading');
+                            .addClass(isPrefix + 'loading');
                     },
                     function(response) {
                         this.cache[url] = true;
 
                         section.html(response)
-                            .removeClass(Toolkit.options.isPrefix + 'loading');
+                            .removeClass(isPrefix + 'loading');
 
                         this.fireEvent('load', response);
                     },
                     function() {
                         section.html(this._errorTemplate())
-                            .removeClass(Toolkit.options.isPrefix + 'loading')
+                            .removeClass(isPrefix + 'loading')
                             .addClass(Toolkit.options.hasPrefix + 'failed');
                     }
                 );
             }
 
             // Toggle tabs
-            this.nav.find('ul > li').removeClass(Toolkit.options.isPrefix + 'active');
+            this.nav.find('ul > li').removeClass(isPrefix + 'active');
 
             // Toggle sections
             if (index === this.currentIndex && options.collapsible) {
@@ -133,13 +134,13 @@
                     section.conceal();
 
                 } else {
-                    tab.parent().addClass(Toolkit.options.isPrefix + 'active');
+                    tab.parent().addClass(isPrefix + 'active');
                     section.reveal();
                 }
             } else {
                 this.hide();
 
-                tab.parent().addClass(Toolkit.options.isPrefix + 'active');
+                tab.parent().addClass(isPrefix + 'active');
                 section.reveal();
             }
 
@@ -167,7 +168,7 @@
          * @param {jQuery.Event} e
          */
         __show: function(e) {
-            if (this.options.preventDefault || (this.options.ajax && $(e.target).attr('href').substr(0, 1) !== '#')) {
+            if (this.options.preventDefault || (this.options.ajax && e.target.href.substr(0, 1) !== '#')) {
                 e.preventDefault();
             }
 
@@ -188,7 +189,7 @@
         cookieDuration: 30,
         getUrl: 'href',
         navElement: '.tabs-nav',
-        sectionsElement: '.tabs-section'
+        sectionElement: '.tabs-section'
     });
 
     /**